requirements.txt
streamlit==1.28.1
python-docx==0.8.11
requests==2.31.0
pydantic==2.5.0
PyPDF2==3.0.1
